add cycle_counter.cpp
