+++ src/include/duckdb/common/cycle_counter.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/common/profiler.hpp
+// duckdb/common/cycle_counter.hpp
 //
 //
 //===----------------------------------------------------------------------===//
@@ -13,9 +13,8 @@
 
 namespace duckdb {
 
-//! The profiler can be used to measure elapsed time
-template <typename T>
-class Profiler {
+//! The cycle counter can be used to measure elapsed cycles
+class CycleCounter {
 public:
 	//! Starts the timer
 	void Start() {
@@ -28,20 +27,14 @@ public:
 		finished = true;
 	}
 
-	//! Returns the elapsed time in seconds. If End() has been called, returns
-	//! the total elapsed time. Otherwise returns how far along the timer is
-	//! right now.
-	double Elapsed() const {
-		auto _end = finished ? end : Tick();
-		return std::chrono::duration_cast<std::chrono::duration<double>>(_end - start).count();
+	uint64_t Elapsed() const {
+		return end - start;
 	}
 
 private:
-	time_point<T> Tick() const {
-		return T::now();
-	}
-	time_point<T> start;
-	time_point<T> end;
+	uint64_t Tick() const;
+	uint64_t start;
+	uint64_t end;
 	bool finished = false;
 };
